Moved tasks.py, views.py and stopping_criteria.py to root and cleaned imports
