+++ smartlearner/__init__.py
@@ -4,7 +4,3 @@ from .interfaces.loss import Loss
 from .interfaces.task import Task
 
 from .trainer import Trainer
-
-from .tasks import tasks
-from .tasks import views
-from .tasks import stopping_criteria
