+++ smartlearner/direction_modifiers/decreasing_learning_rate.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 import numpy as np
 
 from ..utils import sharedX
-from ..interfaces.direction_modifier import DirectionModifier
+from ..interfaces import DirectionModifier
 
 
 class DecreasingLearningRate(DirectionModifier):
