+++ smartlearner/interfaces/__init__.py
@@ -1,7 +1,9 @@
 from .batch_scheduler import BatchScheduler
 from .dataset import Dataset
-from .model import Model
 from .loss import Loss
-from .task import Task
 from .direction_modifier import DirectionModifier
 from .direction_modifier import ParamModifier
+from .model import Model
+from .optimizer import Optimizer
+from .task import Task, RecurrentTask
+from .view import View
