+++ smartlearner/losses/classification_losses.py
@@ -1,6 +1,6 @@
-from ..interfaces.loss import Loss
-
 import theano.tensor as T
+
+from ..interfaces import Loss
 
 
 class NegativeLogLikelihood(Loss):
