+++ smartlearner/views.py
@@ -1,31 +1,8 @@
-from abc import ABCMeta, abstractmethod
-from collections import OrderedDict
 import numpy as np
-from theano import tensor as T
 import theano
+from theano import tensor as T
 
-
-class View(object):
-    __metaclass__ = ABCMeta
-
-    def __init__(self):
-        self.updates = OrderedDict()
-        self.value = None
-        self.last_update = -1
-
-    def view(self, status):
-        if self.last_update != status.current_update:
-            self.value = self.update(status)
-            self.last_update = status.current_update
-
-        return self.value
-
-    @abstractmethod
-    def update(self, status):
-        raise NotImplementedError("Subclass of 'View' must implement 'update(status)'.")
-
-    def __str__(self):
-        return "{0}".format(self.value)
+from .interfaces import View
 
 
 class ItemGetter(View):
